Patch all tests
